+++ packages/plugins/Zoom/src/types.ts
@@ -1,3 +1,5 @@
+import { ZoomIcons } from '@polar/lib-custom-types'
+
 export interface ZoomState {
   zoomLevel: number
   maximumZoomLevel: number
@@ -5,6 +7,7 @@ export interface ZoomState {
 }
 
 export interface ZoomGetters extends ZoomState {
+  icons: ZoomIcons
   maximumZoomLevelActive: boolean
   minimumZoomLevelActive: boolean
   renderType: 'iconMenu' | 'independent'