+++ packages/plugins/Attributions/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unpublished
 
+- Feature: Add option to configure used icons.
 - Chore: Upgrade `ol` from `^9.2.4` to `^10.3.1`.
 
 ## 1.3.0