+++ packages/plugins/Legend/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Add option to configure used icons.
+
 ## 1.1.1
 
 - Fix: The Legend plugin will keep working on the error that a layer without entry in the service register has been configured.